Expose underlying palettes
There are some cases arising, particularly for styling of content, etc where we are not dealing with colors in a semantic way, but still we would like the reliability and predictability that goes along with using our pre-defined palettes. To enable this, we will be exposing the palettes as their own variables for direct access. In most cases, the semantic variables should still be used, but some cases (like the previously mentioned content styling) are better served by the underlying palettes.